+++ fearless_simd_tests/tests/generics.rs
@@ -162,7 +162,7 @@ fn generic_array_roundtrip<S: Simd, V: SimdBase<S>>(vector: V) -> V {
 
     let simd = vector.witness();
     let mut vector = vector;
-    let mut array = vector.as_array();
+    let mut array = vector.to_array();
     let array_copy = array;
     #[expect(clippy::clone_on_copy, reason = "Deliberate test")]
     let array_clone = array.clone();
@@ -170,8 +170,8 @@ fn generic_array_roundtrip<S: Simd, V: SimdBase<S>>(vector: V) -> V {
     let _: Option<V::Element> = array_clone.into_iter().next();
     let _: &[V::Element] = array.as_ref();
     let _: &mut [V::Element] = array.as_mut();
-    let _: &V::Array = vector.as_array_ref();
-    let _: &mut V::Array = vector.as_array_mut();
+    let _: &V::Array = vector.as_array();
+    let _: &mut V::Array = vector.as_mut_array();
     let _ = V::load_array_ref(simd, &array_copy);
     vector.store_array(&mut array);
     V::load_array(simd, array)